lms_micro_services/content-service/requirements.txt
# Content Service Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2

# URL Validation & Utilities
requests==2.31.0
validators==0.22.0
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
